- Added Update.Execs option to the updater. Set to false if you don't want to   update executable files. - Updated xDelta3.

+++ trunk/src/pslaunch/binarypatch.cpp
@@ -44,7 +44,7 @@ int PatchFileHelper(FILE* srcFile, FILE* patchFile, FILE* outFile, int bufferSiz
     void* input_Buffer;
     int input_Buffer_Read;
 
-    if ((uint)bufferSize < XD3_ALLOCSIZE)
+    if (bufferSize < XD3_ALLOCSIZE)
         bufferSize = XD3_ALLOCSIZE;
 
     memset (&stream, 0, sizeof (stream));
